pyproject.toml
[project]
name = "thesis-rl"
version = "0.1.0"
description = "Modular MetaDrive RL codebase for thesis experiments"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
  "metadrive-simulator",
  "gymnasium",
  "stable-baselines3==2.2.1",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "numpy>=1.24,<2",
  "torch==2.2.0a0+81ea7a4",
  "tensorboard",
  "rich",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
  "pytest-cov>=5,<6",
  "ruff",
  "mypy",
]

[tool.uv]
package = true

[tool.uv.sources]
metadrive-simulator = { path = "../third-party/metadrive" }
torch = { path = "../third-party/wheels/torch-2.2.0a0+81ea7a4-cp310-cp310-linux_aarch64.whl" }

[tool.uv.extra-build-dependencies]
pandas = ["setuptools<81"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
